generalize MTP daemon to FunctionFS daemon
MTP daemon handling is a hardcoded version of FunctionFS daemon
handling. Allow to specify for each dynamic FunctionFS mode which
daemon command must be executed to start/stop the FunctionFS userspace
daemon for the USB mode and where the FunctionFS device must be mounted
for the daemon. This prepares usb-moded to support more FunctionFS
userspace daemons such as supporting the Android Open Accesory mode.

Signed-off-by: Dylan Van Assche <me@dylanvanassche.be>